+++ tests/src/IntegrationTest/RemoveTest.php
@@ -16,9 +16,11 @@ namespace Rekalogika\Mapper\Tests\IntegrationTest;
 use Rekalogika\Mapper\Tests\Common\FrameworkTestCase;
 use Rekalogika\Mapper\Tests\Fixtures\Remove\MemberDto;
 use Rekalogika\Mapper\Tests\Fixtures\Remove\ObjectWithAdderRemover;
+use Rekalogika\Mapper\Tests\Fixtures\Remove\ObjectWithAdderRemoverWithAllowDeleteAttachedToGetter;
 use Rekalogika\Mapper\Tests\Fixtures\Remove\ObjectWithArray;
 use Rekalogika\Mapper\Tests\Fixtures\Remove\ObjectWithArrayDto;
 use Rekalogika\Mapper\Tests\Fixtures\Remove\ObjectWithArrayDtoWithAllowTargetDeleteAttribute;
+use Rekalogika\Mapper\Tests\Fixtures\Remove\ObjectWithArrayDtoWithAllowTargetDeleteAttributeAttachedToGetter;
 use Rekalogika\Mapper\Tests\Fixtures\Remove\ObjectWithArrayWithoutAllowDeleteAttribute;
 use Rekalogika\Mapper\Tests\Services\Remove\MemberRepository;
 
@@ -146,11 +148,55 @@ class RemoveTest extends FrameworkTestCase
         $this->assertSame($this->repository->get('2'), $objectWithArray->getMembers()[1]);
     }
 
+    public function testRemoveUsingRemoverWithAllowDeleteAttachedToGetter(): void
+    {
+        $objectWithArrayDto = new ObjectWithArrayDto();
+        $objectWithArrayDto->members[] = new MemberDto('1');
+        $objectWithArrayDto->members[] = new MemberDto('2');
+        // 3 is missing, and this should remove 3 from the target object
+
+        $objectWithArray = new ObjectWithAdderRemoverWithAllowDeleteAttachedToGetter();
+        $objectWithArray->addMember($this->repository->get('1'));
+        $objectWithArray->addMember($this->repository->get('2'));
+        $objectWithArray->addMember($this->repository->get('3'));
+        $this->assertCount(3, $objectWithArray->getMembers());
+
+        $this->mapper->map($objectWithArrayDto, $objectWithArray);
+
+        $this->assertCount(2, $objectWithArray->getMembers());
+        $this->assertSame('1', $objectWithArray->getMembers()[0]->getId());
+        $this->assertSame('2', $objectWithArray->getMembers()[1]->getId());
+        $this->assertSame($this->repository->get('1'), $objectWithArray->getMembers()[0]);
+        $this->assertSame($this->repository->get('2'), $objectWithArray->getMembers()[1]);
+    }
+
     public function testRemoveUsingRemoverWithSourceAttribute(): void
     {
         $objectWithArrayDto = new ObjectWithArrayDtoWithAllowTargetDeleteAttribute();
         $objectWithArrayDto->members[] = new MemberDto('1');
         $objectWithArrayDto->members[] = new MemberDto('2');
+        // 3 is missing, and this should remove 3 from the target object
+
+        $objectWithArray = new ObjectWithAdderRemover();
+        $objectWithArray->addMember($this->repository->get('1'));
+        $objectWithArray->addMember($this->repository->get('2'));
+        $objectWithArray->addMember($this->repository->get('3'));
+        $this->assertCount(3, $objectWithArray->getMembers());
+
+        $this->mapper->map($objectWithArrayDto, $objectWithArray);
+
+        $this->assertCount(2, $objectWithArray->getMembers());
+        $this->assertSame('1', $objectWithArray->getMembers()[0]->getId());
+        $this->assertSame('2', $objectWithArray->getMembers()[1]->getId());
+        $this->assertSame($this->repository->get('1'), $objectWithArray->getMembers()[0]);
+        $this->assertSame($this->repository->get('2'), $objectWithArray->getMembers()[1]);
+    }
+
+    public function testRemoveUsingRemoverWithSourceAttributeAttachedToGetter(): void
+    {
+        $objectWithArrayDto = new ObjectWithArrayDtoWithAllowTargetDeleteAttributeAttachedToGetter();
+        $objectWithArrayDto->addMember(new MemberDto('1'));
+        $objectWithArrayDto->addMember(new MemberDto('2'));
         // 3 is missing, and this should remove 3 from the target object
 
         $objectWithArray = new ObjectWithAdderRemover();